backend/requirements-test.txt
# Backend testing dependencies
pytest==8.3.5
pytest-django==4.9.0
pytest-cov==6.0.0
factory-boy==3.3.1
Faker==36.2.2
